fix bugs

+++ reviews_ali/views.py
@@ -76,8 +76,9 @@ class DemonstationView(View):
 
     def get(self, request, *args, **kwargs):
         image_models = ImageAI.objects.all()
-        for image in image_models:
-            image.delete()
+        if image_models:
+            for image in image_models:
+                image.delete()
         return render(request, 'reviews_ali/video.html', {})
 
 class AIView(View):
